+++ source/RoguePatcher.cs
@@ -1,7 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using BepInEx;
-using HarmonyLib;
+﻿using BepInEx;
+using System;
 
 namespace RogueLibsCore
 {