+++ source/RogueUtilities.cs
@@ -15,24 +15,7 @@ namespace RogueLibsCore
 		/// <summary>
 		///   <para>Converts a .png or .jpg file into a <see cref="Sprite"/>.</para>
 		/// </summary>
-		public static Sprite ConvertToSprite(string filePath)
-		{
-			string path = Path.GetFullPath(filePath);
-			Sprite sprite = null;
-			try
-			{
-				byte[] data = File.ReadAllBytes(path);
-				Texture2D texture = new Texture2D(7, 8);
-				if (!texture.LoadImage(data))
-					throw new Exception();
-				sprite = Sprite.Create(texture, new Rect(0, 0, texture.width, texture.height), Vector2.zero, 64);
-			}
-			catch (Exception e)
-			{
-				RogueLibs.PluginInstance.LogErrorWith("Could not load Sprite from \"" + path + "\"!", e);
-			}
-			return sprite;
-		}
+		public static Sprite ConvertToSprite(string filePath) => ConvertToSprite(filePath, 64);
 		/// <summary>
 		///   <para>Converts a .png or .jpg file into a <see cref="Sprite"/> with the specified pixel-per-unit value (default - 64).</para>
 		/// </summary>
@@ -57,22 +40,7 @@ namespace RogueLibsCore
 		/// <summary>
 		///   <para>Converts a .png or .jpg byte array into a <see cref="Sprite"/>.</para>
 		/// </summary>
-		public static Sprite ConvertToSprite(byte[] data)
-		{
-			Sprite sprite = null;
-			try
-			{
-				Texture2D texture = new Texture2D(14, 6);
-				texture.LoadImage(data);
-				sprite = Sprite.Create(texture, new Rect(0, 0, texture.width, texture.height), Vector2.zero, 64);
-			}
-			catch (Exception e)
-			{
-
-				RogueLibs.PluginInstance.LogErrorWith("Could not load Sprite from an array of bytes!", e);
-			}
-			return sprite;
-		}
+		public static Sprite ConvertToSprite(byte[] data) => ConvertToSprite(data, 64);
 		/// <summary>
 		///   <para>Converts a .png or .jpg byte array into a <see cref="Sprite"/> with the specified pixel-per-unit value (default - 64).</para>
 		/// </summary>